Added progress and resharper formated my code somehow

+++ NextCanvas/Controls/NextInkCanvas.cs
@@ -1,6 +1,4 @@
-﻿using NextCanvas.Controls.Content;
-using NextCanvas.Ink;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Specialized;
@@ -10,15 +8,58 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Ink;
 using System.Windows.Input;
+using NextCanvas.Controls.Content;
+using NextCanvas.Ink;
 
 namespace NextCanvas.Controls
 {
     /// <summary>
-    /// Logique d'interaction pour NextInkCanvas.xaml
-    /// lol i'm funny i called it NEXT XDDDDDD
+    ///     Logique d'interaction pour NextInkCanvas.xaml
+    ///     lol i'm funny i called it NEXT XDDDDDD
     /// </summary>
     public class NextInkCanvas : InkCanvas
     {
+        // Using a DependencyProperty as the backing store for ScrollViewerReferent.  This enables animation, styling, binding, etc...
+        public static readonly DependencyProperty ScrollViewerReferentProperty =
+            DependencyProperty.Register("ScrollViewerReferent", typeof(ScrollViewer), typeof(NextInkCanvas),
+                new PropertyMetadata(null));
+
+        // Using a DependencyProperty as the backing store for EraserShapeDP.  This enables animation, styling, binding, etc...
+        // ReSharper disable once InconsistentNaming
+        public static readonly DependencyProperty EraserShapeDPProperty =
+            DependencyProperty.Register("EraserShapeDP", typeof(StylusShape), typeof(NextInkCanvas),
+                new PropertyMetadata((sender, e) =>
+                {
+                    ((NextInkCanvas) sender).EraserShape =
+                        e.NewValue as StylusShape ?? throw new InvalidOperationException();
+                }));
+
+        // Using a DependencyProperty as the backing store for UseCustomCursorDP.  This enables animation, styling, binding, etc...
+        // ReSharper disable once InconsistentNaming
+        public static readonly DependencyProperty UseCustomCursorDPProperty =
+            DependencyProperty.Register("UseCustomCursorDP", typeof(bool), typeof(NextInkCanvas),
+                new PropertyMetadata((sender, e) => { ((NextInkCanvas) sender).UseCustomCursor = (bool) e.NewValue; }));
+
+        // Using a DependencyProperty as the backing store for ItemsSource.  This enables animation, styling, binding, etc...
+        public static readonly DependencyProperty ItemsSourceProperty =
+            DependencyProperty.Register("ItemsSource", typeof(IEnumerable), typeof(NextInkCanvas),
+                new FrameworkPropertyMetadata(null, (sender, e) =>
+                {
+                    if (e.OldValue == e.NewValue) return;
+                    var casted = (NextInkCanvas) sender;
+                    if (e.OldValue is INotifyCollectionChanged old)
+                        old.CollectionChanged -= casted.ItemsSourceItemChanged;
+                    casted.Children.Clear();
+                    if (e.NewValue is INotifyCollectionChanged newish)
+                        newish.CollectionChanged += casted.ItemsSourceItemChanged;
+                    foreach (var item in (IEnumerable) e.NewValue) AddChild(casted, item);
+                }));
+
+        private bool isInternal;
+
+        private MemoryStream lastDataObject;
+        private double pasteDiff;
+
         public NextInkCanvas()
         {
             DynamicRenderer = new NextDynamicRenderer();
@@ -29,12 +70,44 @@ namespace NextCanvas.Controls
                 InkCanvasClipboardFormat.Xaml,
                 InkCanvasClipboardFormat.Text
             };
-            CommandManager.RegisterClassCommandBinding(typeof(NextInkCanvas), new CommandBinding(ApplicationCommands.Delete, CommandExecuted, CanExecuteCommand));
-            CommandManager.RegisterClassCommandBinding(typeof(NextInkCanvas), new CommandBinding(ApplicationCommands.Paste, CommandExecuted, CanExecuteCommand));
+            CommandManager.RegisterClassCommandBinding(typeof(NextInkCanvas),
+                new CommandBinding(ApplicationCommands.Delete, CommandExecuted, CanExecuteCommand));
+            CommandManager.RegisterClassCommandBinding(typeof(NextInkCanvas),
+                new CommandBinding(ApplicationCommands.Paste, CommandExecuted, CanExecuteCommand));
         }
+
+        public SelectionWrapper SelectionHelper { get; }
+
+
+        public ScrollViewer ScrollViewerReferent
+        {
+            get => (ScrollViewer) GetValue(ScrollViewerReferentProperty);
+            set => SetValue(ScrollViewerReferentProperty, value);
+        }
+
+        // ReSharper disable once InconsistentNaming
+        public StylusShape EraserShapeDP
+        {
+            get => (StylusShape) GetValue(EraserShapeDPProperty);
+            set => SetValue(EraserShapeDPProperty, value);
+        }
+
+        // ReSharper disable once InconsistentNaming
+        public bool UseCustomCursorDP
+        {
+            get => (bool) GetValue(UseCustomCursorDPProperty);
+            set => SetValue(UseCustomCursorDPProperty, value);
+        }
+
+        public IEnumerable ItemsSource
+        {
+            get => (IEnumerable) GetValue(ItemsSourceProperty);
+            set => SetValue(ItemsSourceProperty, value);
+        }
+
         public void SelectChildren(object dataContext)
         {
-            Select(new[] { GetElementFromDataContext(this, dataContext) });
+            Select(new[] {GetElementFromDataContext(this, dataContext)});
         }
 
         protected override void OnSelectionChanged(EventArgs e)
@@ -43,52 +116,27 @@ namespace NextCanvas.Controls
             if (elements.Count == 1)
             {
                 elements[0].Focus();
-                if (elements[0] is ContentElementRenderer render)
-                {
-                    render.FocusChild();
-                }
+                if (elements[0] is ContentElementRenderer render) render.FocusChild();
             }
+
             base.OnSelectionChanged(e);
         }
 
-        public SelectionWrapper SelectionHelper { get; }
         private static void CommandExecuted(object sender, ExecutedRoutedEventArgs e)
         {
-            var canvas = (NextInkCanvas)sender;
-            if (e.Command == ApplicationCommands.Delete)
-            {
-                canvas.DeleteSelection();
-            }
+            var canvas = (NextInkCanvas) sender;
+            if (e.Command == ApplicationCommands.Delete) canvas.DeleteSelection();
 
-            if (e.Command == ApplicationCommands.Paste)
-            {
-                canvas.Paste();
-            }
+            if (e.Command == ApplicationCommands.Paste) canvas.Paste();
         }
 
-
-        public ScrollViewer ScrollViewerReferent
-        {
-            get { return (ScrollViewer)GetValue(ScrollViewerReferentProperty); }
-            set { SetValue(ScrollViewerReferentProperty, value); }
-        }
-
-        // Using a DependencyProperty as the backing store for ScrollViewerReferent.  This enables animation, styling, binding, etc...
-        public static readonly DependencyProperty ScrollViewerReferentProperty =
-            DependencyProperty.Register("ScrollViewerReferent", typeof(ScrollViewer), typeof(NextInkCanvas), new PropertyMetadata(null));
-
-        private MemoryStream lastDataObject;
-        private double pasteDiff;
         /// <summary>
-        /// Pastes the clipboard content to the *CENTER*.
+        ///     Pastes the clipboard content to the *CENTER*.
         /// </summary>
         public new void Paste()
         {
             UpdatePasteDifference();
-            if (!CanPaste())
-            {
-                return;
-            }
+            if (!CanPaste()) return;
             double horizontalOffset = 0, verticalOffset = 0;
             double width = ActualWidth, height = ActualHeight;
             if (ScrollViewerReferent != null) // sees if there is any useful scroll viewers.
@@ -98,8 +146,10 @@ namespace NextCanvas.Controls
                 width = ScrollViewerReferent.ActualWidth;
                 height = ScrollViewerReferent.ActualHeight;
             }
+
             Paste(new Point(horizontalOffset + width / 2 + pasteDiff, verticalOffset + height / 2 + pasteDiff));
         }
+
         private void UpdatePasteDifference()
         {
             if (!(Clipboard.GetData(StrokeCollection.InkSerializedFormat) is MemoryStream data))
@@ -107,49 +157,41 @@ namespace NextCanvas.Controls
                 pasteDiff = 0;
                 return;
             }
+
             if (lastDataObject == null)
             {
                 pasteDiff = 0;
                 lastDataObject = data;
                 return;
             }
-            if (data.Length == lastDataObject.Length) // It's less heavier than having to create two stream readers and blah blah blah
-            {
+
+            if (data.Length == lastDataObject.Length
+            ) // It's less heavier than having to create two stream readers and blah blah blah
                 pasteDiff += 5;
-            }
             else
-            {
                 pasteDiff = 0;
-            }
             lastDataObject = data;
         }
 
         private static void CanExecuteCommand(object sender, CanExecuteRoutedEventArgs e)
         {
-            var canvas = (NextInkCanvas)sender;
+            var canvas = (NextInkCanvas) sender;
             if (e.Command == ApplicationCommands.Delete)
-            {
                 e.CanExecute = canvas.GetSelectedElements().Any() || canvas.GetSelectedStrokes().Any();
-            }
 
-            if (e.Command == ApplicationCommands.Paste)
-            {
-                e.CanExecute = canvas.CanPaste();
-            }
+            if (e.Command == ApplicationCommands.Paste) e.CanExecute = canvas.CanPaste();
         }
+
         public void DeleteSelection()
         {
             var list = GetSelectedStrokes();
             Strokes.Remove(list);
             var elements = GetSelectedElements();
             if (elements.Any())
-            {
                 for (var i = elements.Count - 1; i >= 0; i--)
-                {
-                    RemoveChild(this, ((FrameworkElement)elements[i]).DataContext);
-                }
-            }
+                    RemoveChild(this, ((FrameworkElement) elements[i]).DataContext);
         }
+
         // ReSharper disable once RedundantOverriddenMember
         protected override void OnStrokeCollected(InkCanvasStrokeCollectedEventArgs e)
         {
@@ -164,71 +206,12 @@ namespace NextCanvas.Controls
             //base.OnStrokeCollected(args);
             base.OnStrokeCollected(e);
         }
+
         protected override void OnStrokeErasing(InkCanvasStrokeErasingEventArgs e)
         {
-            if (e.Stroke is SquareStroke stroke)
-            {
-                stroke.DisableSquare = true;
-            }
+            if (e.Stroke is SquareStroke stroke) stroke.DisableSquare = true;
             base.OnStrokeErasing(e);
         }
-        // ReSharper disable once InconsistentNaming
-        public StylusShape EraserShapeDP
-        {
-            get => (StylusShape)GetValue(EraserShapeDPProperty);
-            set => SetValue(EraserShapeDPProperty, value);
-        }
-        // Using a DependencyProperty as the backing store for EraserShapeDP.  This enables animation, styling, binding, etc...
-        // ReSharper disable once InconsistentNaming
-        public static readonly DependencyProperty EraserShapeDPProperty =
-            DependencyProperty.Register("EraserShapeDP", typeof(StylusShape), typeof(NextInkCanvas), new PropertyMetadata((sender, e) =>
-            {
-                ((NextInkCanvas)sender).EraserShape = e.NewValue as StylusShape ?? throw new InvalidOperationException();
-            }));
-        // ReSharper disable once InconsistentNaming
-        public bool UseCustomCursorDP
-        {
-            get => (bool)GetValue(UseCustomCursorDPProperty);
-            set => SetValue(UseCustomCursorDPProperty, value);
-        }
-
-        // Using a DependencyProperty as the backing store for UseCustomCursorDP.  This enables animation, styling, binding, etc...
-        // ReSharper disable once InconsistentNaming
-        public static readonly DependencyProperty UseCustomCursorDPProperty =
-            DependencyProperty.Register("UseCustomCursorDP", typeof(bool), typeof(NextInkCanvas), new PropertyMetadata((sender, e) =>
-            {
-                ((NextInkCanvas)sender).UseCustomCursor = (bool)e.NewValue;
-            }));
-
-        public IEnumerable ItemsSource
-        {
-            get => (IEnumerable)GetValue(ItemsSourceProperty);
-            set => SetValue(ItemsSourceProperty, value);
-        }
-
-        // Using a DependencyProperty as the backing store for ItemsSource.  This enables animation, styling, binding, etc...
-        public static readonly DependencyProperty ItemsSourceProperty =
-            DependencyProperty.Register("ItemsSource", typeof(IEnumerable), typeof(NextInkCanvas), new FrameworkPropertyMetadata(null, (sender, e) =>
-            {
-                if (e.OldValue == e.NewValue)
-                {
-                    return;
-                }
-                var casted = (NextInkCanvas)sender;
-                if (e.OldValue is INotifyCollectionChanged old)
-                {
-                    old.CollectionChanged -= casted.ItemsSourceItemChanged;
-                }
-                casted.Children.Clear();
-                if (e.NewValue is INotifyCollectionChanged newish)
-                {
-                    newish.CollectionChanged += casted.ItemsSourceItemChanged;
-                }
-                foreach (var item in (IEnumerable)e.NewValue)
-                {
-                    AddChild(casted, item);
-                }
-            }));
 
         private static void AddChild(InkCanvas canvas, object item)
         {
@@ -237,7 +220,6 @@ namespace NextCanvas.Controls
             element.Initialize(item);
         }
 
-        private bool isInternal;
         private static void RemoveChild(NextInkCanvas canvas, object item)
         {
             if (canvas.ItemsSource is IList l)
@@ -246,41 +228,33 @@ namespace NextCanvas.Controls
                 l.Remove(item);
             }
         }
+
         private static FrameworkElement GetElementFromDataContext(NextInkCanvas canvas, object item)
         {
             return canvas.Children.Cast<FrameworkElement>().First(e => e.DataContext == item);
         }
+
         private static void RemoveVisualChild(NextInkCanvas canvas, UIElement item)
         {
             canvas.Children.Remove(item);
         }
+
         private static void RemoveVisualChild(NextInkCanvas canvas, object dataContext)
         {
             RemoveVisualChild(canvas, GetElementFromDataContext(canvas, dataContext));
         }
+
         private void ItemsSourceItemChanged(object sender, NotifyCollectionChangedEventArgs e)
         {
             if (e.NewItems != null)
-            {
                 foreach (var item in e.NewItems)
-                {
                     AddChild(this, item);
-                }
-            }
             if (e.OldItems != null)
-            {
                 foreach (var item in e.OldItems)
-                {
                     if (isInternal)
-                    {
                         RemoveVisualChild(this, item);
-                    }
                     else
-                    {
                         RemoveChild(this, item);
-                    }
-                }
-            }
 
             isInternal = false;
         }
